# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Lumen API","description":"Laravel Lumen API Swagger UI with Passport","contact":{"name":"Mitesh Kadivar","email":"kadivarmitesh1021@gmail.com"},"version":"1.0.0"},"paths":{"\/api\/login":{"post":{"tags":["Login"],"summary":"Login","operationId":"login","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application\/json":{}}},"401":{"description":"Unauthenticated"},"400":{"description":"Bad Request"},"404":{"description":"not found"},"403":{"description":"Forbidden"}}}},"\/api\/register":{"post":{"tags":["Register"],"summary":"Register","operationId":"register","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Success","content":{"application\/json":{}}},"401":{"description":"Unauthenticated"},"400":{"description":"Bad Request"},"404":{"description":"not found"},"403":{"description":"Forbidden"}}}},"\/api\/products\/":{"get":{"tags":["Products"],"summary":"Get list of products","description":"Returns list of products","operationId":"getProductList","parameters":[{"name":"page_no","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful operation","content":{"application\/json":{}}},"401":{"description":"Unauthenticated"},"403":{"description":"Forbidden"},"400":{"description":"Bad Request"},"404":{"description":"not found"}},"security":[{"passport":[]}]}}},"components":{"securitySchemes":{"passport":{"type":"oauth2","description":"Laravel passport oauth2 security.","in":"header","scheme":"https","flows":{"password":{"authorizationUrl":"http:\/\/localhost.lumenapi.com\/v1\/oauth\/token","tokenUrl":"http:\/\/localhost.lumenapi.com\/v1\/oauth\/token","refreshUrl":"http:\/\/localhost.lumenapi.com\/v1\/oauth\/token\/refresh","scopes":[]}}}}}}